feat: move location provider to common
