+++ app/src/main/java/org/dhis2/AppComponent.java
@@ -5,8 +5,8 @@ import org.dhis2.commons.network.NetworkUtils;
 import org.dhis2.commons.network.NetworkUtilsModule;
 import org.dhis2.data.dispatcher.DispatcherModule;
 import org.dhis2.data.forms.dataentry.validation.ValidatorModule;
-import org.dhis2.data.location.LocationModule;
-import org.dhis2.data.location.LocationProvider;
+import org.dhis2.commons.locationprovider.LocationModule;
+import org.dhis2.commons.locationprovider.LocationProvider;
 import org.dhis2.commons.prefs.PreferenceModule;
 import org.dhis2.commons.prefs.PreferenceProvider;
 import org.dhis2.commons.schedulers.SchedulerModule;
