+++ app/src/main/java/org/dhis2/data/forms/dataentry/FormView.kt
@@ -43,8 +43,8 @@ import org.dhis2.commons.dialogs.calendarpicker.OnDatePickerListener
 import org.dhis2.commons.extensions.closeKeyboard
 import org.dhis2.commons.extensions.truncate
 import org.dhis2.data.forms.ScanContract
-import org.dhis2.data.location.LocationProvider
-import org.dhis2.data.location.LocationSettingLauncher
+import org.dhis2.commons.locationprovider.LocationProvider
+import org.dhis2.commons.locationprovider.LocationSettingLauncher
 import org.dhis2.databinding.ViewFormBinding
 import org.dhis2.form.Injector
 import org.dhis2.form.data.DataIntegrityCheckResult
