+++ app/src/main/java/org/dhis2/data/forms/dataentry/tablefields/coordinate/CoordinatesView.java
@@ -26,10 +26,9 @@ import androidx.fragment.app.FragmentActivity;
 import com.google.android.material.textfield.TextInputEditText;
 import com.google.android.material.textfield.TextInputLayout;
 
-import org.dhis2.App;
 import org.dhis2.Bindings.StringExtensionsKt;
 import org.dhis2.R;
-import org.dhis2.data.location.LocationSettingLauncher;
+import org.dhis2.commons.locationprovider.LocationSettingLauncher;
 import org.dhis2.maps.geometry.LngLatValidatorKt;
 import org.dhis2.maps.views.MapSelectorActivity;
 import org.dhis2.commons.dialogs.CustomDialog;
