+++ app/src/main/java/org/dhis2/usescases/general/FragmentGlobalAbstract.java
@@ -13,7 +13,7 @@ import androidx.core.app.ActivityOptionsCompat;
 import androidx.fragment.app.Fragment;
 
 import org.dhis2.utils.granularsync.SyncStatusDialog;
-import org.dhis2.data.location.LocationProvider;
+import org.dhis2.commons.locationprovider.LocationProvider;
 import org.dhis2.utils.OnDialogClickListener;
 import org.dhis2.utils.analytics.AnalyticsHelper;
