+++ app/src/main/java/org/dhis2/usescases/searchTrackEntity/mapView/SearchTEMap.kt
@@ -17,7 +17,7 @@ import org.dhis2.animations.CarouselViewAnimations
 import org.dhis2.commons.bindings.clipWithRoundedCorners
 import org.dhis2.commons.data.RelationshipOwnerType
 import org.dhis2.commons.resources.ColorUtils
-import org.dhis2.data.location.LocationSettingLauncher
+import org.dhis2.commons.locationprovider.LocationSettingLauncher
 import org.dhis2.databinding.FragmentSearchMapBinding
 import org.dhis2.maps.ExternalMapNavigation
 import org.dhis2.maps.carousel.CarouselAdapter
