+++ app/src/main/res/values/styles.xml
@@ -123,12 +123,6 @@
         <item name="android:paddingRight">4dp</item>
     </style>
 
-    <style name="CustomDialog" parent="Base.AlertDialog.AppCompat.Light">
-        <item name="android:windowIsFloating">true</item>
-        <item name="android:textColor">?colorPrimary</item>
-        <item name="colorAccent">?colorPrimary</item>
-    </style>
-
     <style name="ButtonRoundedGrey" parent="Base.TextAppearance.AppCompat.Button">
         <item name="android:layout_height">wrap_content</item>
         <item name="android:layout_width">wrap_content</item>
