+++ app/src/test/java/org/dhis2/usescases/eventsWithoutRegistration/eventDetails/EventDetailsIntegrationTest.kt
@@ -11,7 +11,7 @@ import org.dhis2.commons.data.EventCreationType
 import org.dhis2.commons.prefs.PreferenceProvider
 import org.dhis2.commons.resources.ResourceManager
 import org.dhis2.data.dhislogic.DhisPeriodUtils
-import org.dhis2.data.location.LocationProvider
+import org.dhis2.commons.locationprovider.LocationProvider
 import org.dhis2.form.data.GeometryController
 import org.dhis2.form.data.GeometryParserImpl
 import org.dhis2.form.model.FieldUiModel
