+++ commons/src/main/java/org/dhis2/commons/locationprovider/LocationProvider.kt
@@ -1,4 +1,4 @@
-package org.dhis2.data.location
+package org.dhis2.commons.locationprovider
 
 import android.location.Location
