+++ commons/src/main/res/values-vi/strings.xml
@@ -205,5 +205,6 @@
     <string name="error_org_unit_scope">Tìm kiếm trực tuyến chỉ hiển thị trong các đơn vị mà tài khoảng của bạn được cấp.</string>
     <string name="error_invalid_characters">Có các ký tự không hợp lệ trong yêu cầu. Vui lòng liên hệ quản lý để xác nhận cấu hình máy chủ (relaxQueryChars,...)</string>
     <string name="unsupported_field_description">Trường này không hiển thị để nhập dữ liệu bằng điện thoại</string>
+    <string name="enable_location_message">Tính năng chia sẻ vị trí thiết bị của bạn bị tắt. Vui lòng mở nó để thu thập tọa độ.</string>
     <string name="add_image">Thêm hình ảnh</string>
 </resources>
