+++ commons/src/main/res/values/strings.xml
@@ -208,4 +208,6 @@
     <string name="error_invalid_characters">There are invalid characters in the query. Please ask the administrator to verify the server configuration (relaxedQueryChars,...).</string>
     <string name="unsupported_field_description">This field is not available for mobile capture</string>
     <string name="add_image">Add Image</string>
+    <string name="action_ok">Ok</string>
+    <string name="enable_location_message">Your device location is disconnected. Please enable it to capture coordinates</string>
 </resources>
